+++ api/adapters/aws_adapter.py
@@ -1,25 +1,22 @@
 # Copyright 2019
 # All rights reserved. Mihail Butnaru
-""" Enables to create a connection to access
-AWS services.
+""" 
+Creates a connection to access AWS services.
 """
 import logging
 import boto3
-from api.config import Config
 
-class AWSConnection():
-    """ Easy use connection through the boto3 to access
-    AWS Services.
+def init_aws_connection(config):
     """
-    def __init__(self):
-        """ Boto3 client connection based on the configuration.
-            Args:
-                config (str): config parameters
-        """
-        self.config = Config()
-        self.client = boto3.client(
-            self.config.AWS_SERVICE,
-            aws_access_key_id=self.config.AWS_ACCESS_KEY_ID,
-            aws_secret_access_key=self.config.AWS_SECRET_ACCESS_KEY,
-            region_name=self.config.AWS_REGION_NAME
-        )
+    AWS connection through Boto3 to access AWS Services.
+        Args:
+            config (str): config parameters
+    """
+    _log = logging.getLogger(__name__)
+
+    return boto3.client(
+        config.AWS_SERVICE,
+        aws_access_key_id=config.AWS_ACCESS_KEY_ID,
+        aws_secret_access_key=config.AWS_SECRET_ACCESS_KEY,
+        region_name=config.AWS_REGION_NAME
+    )
