+++ api/app.py
@@ -3,12 +3,13 @@
 import logging
 from api.routes.restplus import api
 from flask import Flask, Blueprint
-from api.resources.users import ns as users_namespace
-from api.resources.app_clients import ns as appclients_namespace
-from api.resources.scopes import ns as scopes_namespace
+from api.resources.users_resource import ns as users_namespace
+from api.resources.app_clients_resource import ns as appclients_namespace
+from api.resources.scopes_resource import ns as scopes_namespace
 
 def create_app(config):
-    """ Flask Initialization application
+    """ 
+    Flask Initialization application
     """
     logging.getLogger(__name__)
